+++ Tests/LegacyUnit/BackEnd/ConfirmEventMailFormTest.php
@@ -19,12 +19,12 @@ final class ConfirmEventMailFormTest extends TestCase
     /**
      * @var ConfirmEventMailForm
      */
-    private $subject = null;
+    private $subject;
 
     /**
      * @var TestingFramework
      */
-    private $testingFramework = null;
+    private $testingFramework;
 
     protected function setUp(): void
     {