+++ Tests/LegacyUnit/BackEnd/FlexFormsTest.php
@@ -13,12 +13,12 @@ final class FlexFormsTest extends TestCase
     /**
      * @var FlexForms
      */
-    private $subject = null;
+    private $subject;
 
     /**
      * @var TestingFramework
      */
-    private $testingFramework = null;
+    private $testingFramework;
 
     protected function setUp(): void
     {