+++ Tests/LegacyUnit/BackEnd/SpeakersListTest.php
@@ -23,17 +23,17 @@ final class SpeakersListTest extends TestCase
     /**
      * @var SpeakersList
      */
-    private $subject = null;
+    private $subject;
 
     /**
      * @var TestingFramework
      */
-    private $testingFramework = null;
+    private $testingFramework;
 
     /**
      * @var int PID of a dummy system folder
      */
-    private $dummySysFolderPid = 0;
+    private $dummySysFolderPid;
 
     protected function setUp(): void
     {